Bug Fix for iPhone 14/15 Max
Back button on modal was too high for some devices; this change put the back button at the same height as the back button for the movie detail page.

Also updated the version # for the Apple Store Connect review to show them it was a new release to fix some issues they ran into.