Version 2.2.9(a)
Version 2.2.9 almost complete. Added Favorites page to pull saved movies from the Movie Details page. Still in process for cleaning up the Favorites page, and will then have to add a check on the Movie Detail page to see if the movie has been favorited before, and if it has, set the heart to filled in status of true