Version 2.2.9(b)
Completed the Movie Details page to be able to check whether or not a full heart should be displayed for the movie to indicate it is saved under favorites, and to be able to click (save) and unclick (remove) the heart. Had problem with understanding what to use for the useEffect for running the check for the heart status so that the checkHeart method would run too many times.