Version 2.2.11
Updated the App Settings page to clear the correct key for deleting stored movie data and removed unused code. Updated the Favorites page to immediately  reflect when the the movie data is cleared from the app settings page vs having to shutdown the app and opening it again. Added Close buttons for the Begin and End Date picker on the Search by Date page